+++ packages/anvil-embed-frame/CHANGELOG.md
@@ -2,6 +2,52 @@
 
 All notable changes to this project will be documented in this file.
 
+## [@anvilco/anvil-embed-frame@2.2.1](https://github.com/anvilco/react-ui/compare/@anvilco/anvil-embed-frame@2.2.0...@anvilco/anvil-embed-frame@2.2.1) - 2025-10-27
+
+### Merged
+
+- iframeURL instead of signURL [`#25`](https://github.com/anvilco/react-ui/pull/25)
+
+### Commits
+
+- fix up packaging [`74ce1cb`](https://github.com/anvilco/react-ui/commit/74ce1cb7a77295545ef89b6f43777e874ec5ce60)
+
+## [@anvilco/anvil-embed-frame@2.2.0](https://github.com/anvilco/react-ui/compare/v2.1.4...@anvilco/anvil-embed-frame@2.2.0) - 2024-05-02
+
+## [v2.1.4](https://github.com/anvilco/react-ui/compare/v2.1.3...v2.1.4) - 2025-10-28
+
+### Commits
+
+- being extra cautious [`02beb3c`](https://github.com/anvilco/react-ui/commit/02beb3c650dc774a84cf96166e38eaee5f17b0ca)
+
+## [v2.1.3](https://github.com/anvilco/react-ui/compare/v2.1.2...v2.1.3) - 2025-10-28
+
+### Commits
+
+- set document inside component did mount [`5a055ed`](https://github.com/anvilco/react-ui/commit/5a055edfd2d5dcef9d3b6c4b8d6dd4a8fd1def77)
+
+## [v2.1.2](https://github.com/anvilco/react-ui/compare/v2.1.1...v2.1.2) - 2025-10-27
+
+### Merged
+
+- iframeURL instead of signURL [`#25`](https://github.com/anvilco/react-ui/pull/25)
+- Add className prop and type [`#22`](https://github.com/anvilco/react-ui/pull/22)
+- publish better [`#20`](https://github.com/anvilco/react-ui/pull/20)
+
+### Commits
+
+- fix up packaging [`74ce1cb`](https://github.com/anvilco/react-ui/commit/74ce1cb7a77295545ef89b6f43777e874ec5ce60)
+
+## [v2.1.1](https://github.com/anvilco/react-ui/compare/@anvilco/anvil-embed-frame@2.0.0...v2.1.1) - 2024-04-17
+
+### Merged
+
+- Add style prop and add postMessage helper [`#19`](https://github.com/anvilco/react-ui/pull/19)
+
+### Commits
+
+- Update changelogs [`73908a8`](https://github.com/anvilco/react-ui/commit/73908a846d6cf7da180303270a4d726335ff881f)
+
 ## [@anvilco/anvil-embed-frame@2.0.0](https://github.com/anvilco/react-ui/compare/@anvilco/anvil-embed-frame@2.0.0-alpha.0...@anvilco/anvil-embed-frame@2.0.0) - 2023-09-20
 
 ### Merged
@@ -12,27 +58,27 @@ All notable changes to this project will be documented in this file.
 
 - Update readmes [`c932e4c`](https://github.com/anvilco/react-ui/commit/c932e4c5716c0b113f1840baeb0aa9118facdc26)
 
-## [@anvilco/anvil-embed-frame@2.0.0-alpha.0](https://github.com/anvilco/react-ui/compare/@anvilco/anvil-embed-frame@1.1.0...@anvilco/anvil-embed-frame@2.0.0-alpha.0) - 2023-09-18
+## [@anvilco/anvil-embed-frame@2.0.0-alpha.0](https://github.com/anvilco/react-ui/compare/v1.9.2...@anvilco/anvil-embed-frame@2.0.0-alpha.0) - 2023-09-18
+
+## [v1.9.2](https://github.com/anvilco/react-ui/compare/@anvilco/anvil-embed-frame@1.1.0...v1.9.2) - 2025-10-30
 
 ### Merged
 
+- iframeURL instead of signURL [`#25`](https://github.com/anvilco/react-ui/pull/25)
+- Add className prop and type [`#22`](https://github.com/anvilco/react-ui/pull/22)
+- publish better [`#20`](https://github.com/anvilco/react-ui/pull/20)
+- Add style prop and add postMessage helper [`#19`](https://github.com/anvilco/react-ui/pull/19)
+- [Breaking] Update dependencies [`#16`](https://github.com/anvilco/react-ui/pull/16)
 - Update readme to fix import error [`#15`](https://github.com/anvilco/react-ui/pull/15)
 
 ### Commits
 
 - Update deps [`52a73bd`](https://github.com/anvilco/react-ui/commit/52a73bd5f16d5435f80cd8ae76f20b2834807c40)
 
-## [@anvilco/anvil-embed-frame@1.1.0](https://github.com/anvilco/react-ui/compare/v0.0.0...@anvilco/anvil-embed-frame@1.1.0) - 2022-10-05
-
-## v0.0.0 - 2024-05-02
+## @anvilco/anvil-embed-frame@1.1.0 - 2022-10-05
 
 ### Merged
 
-- Add className prop and type [`#22`](https://github.com/anvilco/react-ui/pull/22)
-- publish better [`#20`](https://github.com/anvilco/react-ui/pull/20)
-- Add style prop and add postMessage helper [`#19`](https://github.com/anvilco/react-ui/pull/19)
-- [Breaking] Update dependencies [`#16`](https://github.com/anvilco/react-ui/pull/16)
-- Update readme to fix import error [`#15`](https://github.com/anvilco/react-ui/pull/15)
 - Add Typescript support to `AnvilEmbedFrame` [`#13`](https://github.com/anvilco/react-ui/pull/13)
 - New `@anvilco/anvil-embed-frame` package [`#12`](https://github.com/anvilco/react-ui/pull/12)
 - Update Readme [`#11`](https://github.com/anvilco/react-ui/pull/11)